added testing to UI

+++ AuctionHouse_WebApplication/src/main/java/com/example/auctionhouse_webapplication/model/Report.java
@@ -7,29 +7,30 @@ import jakarta.persistence.Id;
 
 import java.util.Date;
 
+
+
 @Entity
 public class Report {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     private String title;
     private String description;
     private Date reportDate;
-
     private String email;
 
+    // Konstruktory, Gettery i Settery
 
     public Report() {
     }
 
-
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
